Removed old boss randomizer code
